Align identifer characters rule with go/xgo
