+++ spx-gui/src/models/common/asset-name.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { getSoundName, getSpriteName, normalizeAssetName, normalizeGopIdentifierAssetName } from './asset-name'
+import { getSoundName, getSpriteName, normalizeAssetName } from './asset-name'
 import { Project } from '../project'
 import { Sprite } from '../sprite'
 import { Sound } from '../sound'
@@ -19,33 +19,6 @@ describe('normalizeAssetName', () => {
   it('should work well with non-ascii chars', () => {
     expect(normalizeAssetName('中文-1', 'camel')).toBe('中文-1')
     expect(normalizeAssetName('中文 2', 'pascal')).toBe('中文 2')
-  })
-})
-
-describe('normalizeGopIdentifierAssetName', () => {
-  it('should work well with camel case', () => {
-    expect(normalizeGopIdentifierAssetName('abc', 'camel')).toBe('abc')
-    expect(normalizeGopIdentifierAssetName('abc def---ghi__jkl.mno', 'camel')).toBe('abcDefGhiJklMno')
-    expect(normalizeGopIdentifierAssetName('', 'camel')).toBe('')
-    expect(normalizeGopIdentifierAssetName('中文', 'camel')).toBe('')
-    expect(normalizeGopIdentifierAssetName('中文en', 'camel')).toBe('en')
-    expect(normalizeGopIdentifierAssetName('中文En', 'camel')).toBe('en')
-    expect(normalizeGopIdentifierAssetName('abc中 文en', 'camel')).toBe('abcEn')
-    expect(normalizeGopIdentifierAssetName('abc中 文en', 'camel')).toBe('abcEn')
-    expect(normalizeGopIdentifierAssetName('123abc 456', 'camel')).toBe('abc456')
-    expect(normalizeGopIdentifierAssetName(repeat('a', 120), 'camel')).toBe(repeat('a', 20))
-  })
-  it('should work well with pascal case', () => {
-    expect(normalizeGopIdentifierAssetName('abc', 'pascal')).toBe('Abc')
-    expect(normalizeGopIdentifierAssetName('abc def---ghi__jkl.mno', 'pascal')).toBe('AbcDefGhiJklMno')
-    expect(normalizeGopIdentifierAssetName('', 'pascal')).toBe('')
-    expect(normalizeGopIdentifierAssetName('中文', 'pascal')).toBe('')
-    expect(normalizeGopIdentifierAssetName('中文en', 'pascal')).toBe('En')
-    expect(normalizeGopIdentifierAssetName('中文En', 'pascal')).toBe('En')
-    expect(normalizeGopIdentifierAssetName('abc中 文en', 'pascal')).toBe('AbcEn')
-    expect(normalizeGopIdentifierAssetName('abc中 文en', 'pascal')).toBe('AbcEn')
-    expect(normalizeGopIdentifierAssetName('123abc 456', 'pascal')).toBe('Abc456')
-    expect(normalizeGopIdentifierAssetName(repeat('a', 120), 'pascal')).toBe('A' + repeat('a', 19))
   })
 })
 
@@ -91,8 +64,4 @@ describe('getSoundName', () => {
 
 function mockFile() {
   return fromText('', '')
-}
-
-function repeat(str: string, times: number) {
-  return Array.from({ length: times + 1 }).join(str)
 }
